HBX-1606: Move class 'org.hibernate.cfg.binder.MetaAttributesBinder' to 'org.hibernate.tool.internal.reveng.MetaAttributesBinder'
