+++ orm/src/main/java/org/hibernate/cfg/binder/PropertyBinder.java
@@ -4,6 +4,7 @@ import org.hibernate.mapping.Property;
 import org.hibernate.mapping.Table;
 import org.hibernate.mapping.Value;
 import org.hibernate.tool.api.reveng.ReverseEngineeringStrategy;
+import org.hibernate.tool.internal.reveng.MetaAttributesBinder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
